Add files via upload
Merged both scripts into 1, we also no longer need a 'paint' script on each trigger, instead we now have a list of up to 20 triggers in the script properties inspector in editor, and a list of 20 colors that correspond with the list of triggers, in order.  We also have a single slot for a 'rainbow/random colors' and 'cleanser' trigger now too.